+++ src/components/Calendar/CHANGELOG.md
@@ -1,3 +1,4 @@
 - 2025-10-03: Initial version.
 - 2025-10-08: Update its `border-color` and `border-radius` styles.
 - 2025-11-01: Make it use `utils/jss` and put its styles in CSS layer, etc.
+- 2025-11-26: Support the month/year picker.